Clean up Feedback::States namespace and spec
